improved chunk generation

+++ include/chunk.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <memory>
-#include <map>
+#include <unordered_map>
 #include "glm/glm.hpp"
 #include "instance3D.h"
 #include "noise.h"
@@ -8,18 +8,18 @@
 
 namespace Voxel {
     namespace Game {
-        const size_t SIZE = sizeof(uint16_t) * 8;
+        const int SIZE = sizeof(uint16_t) * 8;
 
         class Chunk {
         public:
-            static std::shared_ptr<Chunk> create(Noise& noise, glm::vec3 position);
+            static std::shared_ptr<Chunk> create(Noise& noise, glm::ivec3 position);
 
             Chunk() = default;
-            Chunk(Noise& noise, glm::vec3 position);
+            Chunk(Noise& noise, glm::ivec3 position);
             void build_mesh();
             void render();
 
-            glm::vec3 position;
+            glm::ivec3 position;
             uint16_t voxels[SIZE * SIZE * 3] = {};
         private:
             unsigned int data[SIZE * SIZE * SIZE];
